Moved SQLRepository directory
