+++ src/Spudbot/Repository/SQL/ThreadRepository.php
@@ -7,7 +7,7 @@ use Discord\Parts\Part;
 use Spudbot\Collection;
 use Spudbot\Model;
 use Spudbot\Model\Thread;
-use Spudbot\Repository;
+use Spudbot\Repository\SQLRepository;
 
 class ThreadRepository extends SQLRepository
 {
